Remove DropletFactory
