+++ tests/Functional/Services/MachineProvider/DigitalOcean/DigitalOceanMachineProviderTest.php
@@ -79,6 +79,33 @@ class DigitalOceanMachineProviderTest extends AbstractBaseFunctionalTest
         self::assertSame($ipAddresses, ObjectReflector::getProperty($this->worker, 'ip_addresses'));
     }
 
+    public function testCreateThrowsDropletLimitException(): void
+    {
+        $this->mockHandler->append(
+            new Response(
+                422,
+                [
+                    'content-type' => 'application/json',
+                ],
+                (string) json_encode([
+                    'id' => 'unprocessable_entity',
+                    'message' => 'creating this/these droplet(s) will exceed your droplet limit',
+                ])
+            )
+        );
+
+        try {
+            $this->machineProvider->create($this->worker);
+            self::fail('WorkerApiActionException not thrown');
+        } catch (WorkerApiActionException $workerApiActionException) {
+            self::assertSame($this->worker, $workerApiActionException->getWorker());
+            self::assertEquals(
+                new ValidationFailedException('creating this/these droplet(s) will exceed your droplet limit', 422),
+                $workerApiActionException->getRemoteApiException()
+            );
+        }
+    }
+
     public function testHydrateSuccess(): void
     {
         $remoteId = 123;
