--- tests/Functional/Services/MachineProvider/DigitalOcean/DropletFactoryTest.php
@@ -1,88 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Tests\Functional\Services\MachineProvider\DigitalOcean;
-
-use App\Exception\MachineProvider\WorkerApiActionException;
-use App\Model\ProviderInterface;
-use App\Services\MachineProvider\DigitalOcean\DropletFactory;
-use App\Services\WorkerFactory;
-use App\Tests\AbstractBaseFunctionalTest;
-use App\Tests\Services\HttpResponseFactory;
-use DigitalOceanV2\Entity\Droplet as DropletEntity;
-use DigitalOceanV2\Exception\ValidationFailedException;
-use GuzzleHttp\Handler\MockHandler;
-use GuzzleHttp\Psr7\Response;
-
-class DropletFactoryTest extends AbstractBaseFunctionalTest
-{
-    private DropletFactory $dropletFactory;
-    private WorkerFactory $workerFactory;
-    private MockHandler $mockHandler;
-
-    protected function setUp(): void
-    {
-        parent::setUp();
-
-        $dropletFactory = self::$container->get(DropletFactory::class);
-        if ($dropletFactory instanceof DropletFactory) {
-            $this->dropletFactory = $dropletFactory;
-        }
-
-        $workerFactory = self::$container->get(WorkerFactory::class);
-        if ($workerFactory instanceof WorkerFactory) {
-            $this->workerFactory = $workerFactory;
-        }
-
-        $mockHandler = self::$container->get(MockHandler::class);
-        if ($mockHandler instanceof MockHandler) {
-            $this->mockHandler = $mockHandler;
-        }
-    }
-
-    public function testCreateSuccess(): void
-    {
-        $remoteId = 123;
-        $dropletData = [
-            'id' => $remoteId,
-        ];
-
-        $expectedDropletEntity = new DropletEntity($dropletData);
-        $this->mockHandler->append(HttpResponseFactory::fromDropletEntity($expectedDropletEntity));
-
-        $worker = $this->workerFactory->create(md5('id content'), ProviderInterface::NAME_DIGITALOCEAN);
-        $createDropletEntity = $this->dropletFactory->create($worker);
-
-        self::assertSame($remoteId, $createDropletEntity->id);
-    }
-
-    public function testCreateThrowsDropletLimitException(): void
-    {
-        $this->mockHandler->append(
-            new Response(
-                422,
-                [
-                    'content-type' => 'application/json',
-                ],
-                (string) json_encode([
-                    'id' => 'unprocessable_entity',
-                    'message' => 'creating this/these droplet(s) will exceed your droplet limit',
-                ])
-            )
-        );
-
-        $worker = $this->workerFactory->create(md5('id content'), ProviderInterface::NAME_DIGITALOCEAN);
-
-        try {
-            $this->dropletFactory->create($worker);
-            self::fail('WorkerApiActionException not thrown');
-        } catch (WorkerApiActionException $workerApiActionException) {
-            self::assertSame($worker, $workerApiActionException->getWorker());
-            self::assertEquals(
-                new ValidationFailedException('creating this/these droplet(s) will exceed your droplet limit', 422),
-                $workerApiActionException->getRemoteApiException()
-            );
-        }
-    }
-}
